Move classes in right namespaces
