+++ src/Core/Grid/Column/Type/ColorColumn.php
@@ -26,34 +26,10 @@
 
 namespace PrestaShop\PrestaShop\Core\Grid\Column\Type;
 
-use PrestaShop\PrestaShop\Core\Grid\Column\AbstractColumn;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
 /**
- * Column which background color can be configured.
+ * @deprecated since 8.1 and will be removed in next major.
+ * Use \PrestaShop\PrestaShop\Core\Grid\Column\Type\Common\ColorColumn instead.
  */
-final class ColorColumn extends AbstractColumn
+final class ColorColumn extends \PrestaShop\PrestaShop\Core\Grid\Column\Type\Common\ColorColumn
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getType()
-    {
-        return 'color';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function configureOptions(OptionsResolver $resolver)
-    {
-        parent::configureOptions($resolver);
-
-        $resolver
-            ->setRequired([
-                'field',
-                'color_field',
-            ])
-        ;
-    }
 }
