+++ src/Core/Grid/Column/Type/Common/PreviewColumn.php
@@ -24,12 +24,42 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShop\PrestaShop\Core\Grid\Column\Type;
+namespace PrestaShop\PrestaShop\Core\Grid\Column\Type\Common;
+
+use PrestaShop\PrestaShop\Core\Grid\Column\AbstractColumn;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * @deprecated since 8.1 and will be removed in next major.
- * Use \PrestaShop\PrestaShop\Core\Grid\Column\Type\Common\DataColumn instead.
+ * Allows adding Preview functionality to grid rows
  */
-final class DataColumn extends \PrestaShop\PrestaShop\Core\Grid\Column\Type\Common\DataColumn
+class PreviewColumn extends AbstractColumn
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function getType()
+    {
+        return 'preview';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver
+            ->setRequired([
+                'icon_expand',
+                'icon_collapse',
+                'preview_data_route',
+            ])
+            ->setDefined([
+                'preview_route_params',
+            ])
+            ->setAllowedTypes('preview_data_route', 'string')
+            ->setAllowedTypes('preview_route_params', 'array')
+            ->setAllowedTypes('icon_expand', 'string')
+            ->setAllowedTypes('icon_collapse', 'string')
+        ;
+    }
 }
