+++ src/Core/Grid/Column/Type/OrderPriceColumn.php
@@ -26,36 +26,10 @@
 
 namespace PrestaShop\PrestaShop\Core\Grid\Column\Type;
 
-use PrestaShop\PrestaShop\Core\Grid\Column\AbstractColumn;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
 /**
- * Displays order price
+ * @deprecated since 8.1 and will be removed in next major.
+ * Use \PrestaShop\PrestaShop\Core\Grid\Column\Type\Order\OrderPriceColumn instead.
  */
-final class OrderPriceColumn extends AbstractColumn
+final class OrderPriceColumn extends \PrestaShop\PrestaShop\Core\Grid\Column\Type\Order\OrderPriceColumn
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getType()
-    {
-        return 'order_price';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function configureOptions(OptionsResolver $resolver)
-    {
-        parent::configureOptions($resolver);
-
-        $resolver
-            ->setRequired([
-                'field',
-                'is_paid_field',
-            ])
-            ->setAllowedTypes('field', 'string')
-            ->setAllowedTypes('is_paid_field', 'string')
-        ;
-    }
 }
